# backend/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swings API",
    "description": "Swings public + admin HTTP API. Mutating endpoints are annotated; GETs are still fully callable but not listed here (see FDN-02 scope).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to deploying host"
    }
  ],
  "paths": {
...
    "/api/webhooks/stripe": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "operationId": "stripe_webhook",
        "requestBody": {
          "description": "Raw Stripe webhook JSON payload",
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed"
          },
          "400": {
            "description": "Invalid signature or payload"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}